+++ maskrcnn_benchmark/modeling/rpn/retinanet/inference.py
@@ -1,12 +1,12 @@
 import torch
 
 from maskrcnn_benchmark.modeling.box_coder import BoxCoder
+from maskrcnn_benchmark.modeling.utils import cat
 from maskrcnn_benchmark.structures.bounding_box import BoxList
 from maskrcnn_benchmark.structures.boxlist_ops import cat_boxlist
 from maskrcnn_benchmark.structures.boxlist_ops import boxlist_nms
 from maskrcnn_benchmark.structures.boxlist_ops import remove_small_boxes
 
-from ..utils import cat
 
 
 class RetinaNetPostProcessor(torch.nn.Module):
